Added initial opensearch setup
