+++ .env
@@ -50,3 +50,4 @@ SHUFFLE_CONTAINER_AUTO_CLEANUP=false
 SHUFFLE_OPENSEARCH_URL="http://shuffle-opensearch:9200"
 SHUFFLE_OPENSEARCH_USERNAME=""
 SHUFFLE_OPENSEARCH_PASSWORD=""
+SHUFFLE_ELASTIC="true"
